Fixed a bug on startup for missing source directory Updated version number

+++ StormLoader/StormLoader/mod_handling/ModInstaller.cs
@@ -23,6 +23,11 @@ namespace StormLoader.mod_handling
         }
         public void DeserializeOrCreateMods(string source)
         {
+            if (source == "" || source == null || !Directory.Exists(source))
+            {
+                return;
+            }
+
             if (!File.Exists(source + "/StormLoader_install_info.bin"))
             {
                 File.Create(source + "/StormLoader_install_info.bin");
